Removed random
Having a pop() call and not using it's return value is only getting around half the efficiency of the function itself. Instead of using random to determine the key we want, we remove the nessasity of having random in this script entirely, we remove the choice() call completely, and we get the most value out of the pop() function.