Refactored
So, I just want to start by saying sorry for messing around with this class so much. I think it's a very easy class to understand. However, I would just like to note that this class acts as a psuedo generator in that it returns instead of yielding.

In this commit I removed all internal and local variables that were not used outside of the class. I noticed how self.visited was not actually used anywhere accept as a placeholder to reset the self.keys list. This commit completely does away with self._reset and self.visited by first moving the first element of the keys array to the end, and then returning the "next" key with the pop() functions return call.